Add files via upload
This is an assignment for the Text Analytics class for the Masters in Business Analytics program. Here, we try to evaluate the reviews for Airbnb and Booking.com, and how to create business value from those. It uses web scraping packages, to pull reviews from "Trustpilot", and then do a text analysis to create value.